+++ source/extensions/isaacsim.core.simulation_manager/plugins/isaacsim.core.simulation_manager/UsdNoticeListener.cpp
@@ -124,6 +124,6 @@ bool UsdNoticeListener::isEnabled()
     return this->m_enableFlag;
 }
 
-}
-}
-}
+} // namespace simulation_manager
+} // namespace core
+} // namespace isaacsim